fix bug

+++ app/src/main/java/com/vietinterview/getbee/api/response/listcv/CvList.java
@@ -110,7 +110,7 @@ public class CvList implements Parcelable {
     private Integer fromSalary;
     @SerializedName("fee")
     @Expose
-    private Integer fee;
+    private Long fee;
     @SerializedName("levelName")
     @Expose
     private String levelName;
@@ -207,7 +207,7 @@ public class CvList implements Parcelable {
         if (in.readByte() == 0) {
             fee = null;
         } else {
-            fee = in.readInt();
+            fee = in.readLong();
         }
         levelName = in.readString();
     }
@@ -488,11 +488,11 @@ public class CvList implements Parcelable {
         this.fromSalary = fromSalary;
     }
 
-    public Integer getFee() {
+    public Long getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Long fee) {
         this.fee = fee;
     }
 
@@ -617,7 +617,7 @@ public class CvList implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(fee);
+            parcel.writeLong(fee);
         }
         parcel.writeString(levelName);
     }
